fix: give up with infinite carousel
